+++ Top4Net/Request/ItemVipAddRequest.cs
@@ -35,7 +35,7 @@ namespace Taobao.Top.Api.Request
         public Nullable<DateTime> ListTime { get; set; }
         public string LocationCity { get; set; }
         public string LocationState { get; set; }
-        public Nullable<long> Num { get; set; }
+        public Nullable<int> Num { get; set; }
         public string OuterId { get; set; }
         public string PicPath { get; set; }
         public string PostFee { get; set; }
@@ -53,7 +53,7 @@ namespace Taobao.Top.Api.Request
         public string StuffStatus { get; set; }
         public string Title { get; set; }
         public string Type { get; set; }
-        public Nullable<long> ValidThru { get; set; }
+        public Nullable<int> ValidThru { get; set; }
 
         #region ITopRequest Members
